added phone numbers to Users and editing capabilities
