+++ prisma/schema.prisma
@@ -61,6 +61,7 @@ model User {
   name          String?
   email         String?   @unique
   emailVerified DateTime? @map("email_verified")
+  phone         String?   @unique
   image         String?
   accounts      Account[]
   sessions      Session[]
